requirements.txt
flair==0.13.0
nltk==3.8.1
dspy-ai==2.1.1
pandas==2.1.1
scikit-learn==1.3.2
chardet==5.2.0
sentence-transformers==2.2.2
transformers==4.34.1
scipy==1.10.1
fastchat
wikipedia==1.4.0
Wikipedia-API==0.6.0
rouge-score
toml